auto init apiClient
